datnd update

+++ FDex.Application/DTOs/User/UserDTOLeaderboardItemView.cs
@@ -7,7 +7,7 @@ namespace FDex.Application.DTOs.User
 	{
 		public string Wallet { get; set; }
 		public BigInteger TradingVol { get; set; }
-		public decimal AvgLeverage { get; set; }
+		public double AvgLeverage { get; set; }
 		public int Win { get; set; }
 		public int Loss { get; set; }
 		public BigInteger PNLwFees { get; set; }
